gestion funciona header en todos

+++ src/main/resources/META-INF/resources/js/comercio.js
@@ -50,15 +50,15 @@ Vue.createApp({
         // Mostrar el modal para confirmar la acción de cerrar sesión
         confirmLogout() {
             this.modalMessage = '¿Estás seguro que quieres cerrar sesión?';
-            this.actionToConfirm = this.logoutComercio;  // Establecer la acción que se confirmará
-            this.showModal = true;  // Mostrar el modal
+            this.actionToConfirm = this.logoutComercio();  // Establecer la acción que se confirmará
+            //this.showModal = true;  // Mostrar el modal
         },
 
         // Mostrar el modal para confirmar la eliminación de cuenta
         confirmDeleteAccount() {
             this.modalMessage = '¿Estás seguro que quieres eliminar tu cuenta?';
-            this.actionToConfirm = this.deleteComercio;  // Establecer la acción que se confirmará
-            this.showModal = true;  // Mostrar el modal
+            this.actionToConfirm = this.deleteComercio();  // Establecer la acción que se confirmará
+            //this.showModal = true;  // Mostrar el modal
         },
 
         // Ejecutar la acción confirmada (cerrar sesión o eliminar cuenta)
@@ -91,8 +91,6 @@ Vue.createApp({
                 if (response.status === 200) {
                     console.log(response.data)
                     const tipoComercio = response.data.negocio.tipo_negocio
-
-                    alert("Inicio de sesión exitoso");
 
                     // PEQUEÑO RETRASO para asegurar que la cookie se guarda
                     setTimeout(() => {
@@ -199,44 +197,55 @@ Vue.createApp({
         ,
 
         async deleteComercio() {
+            console.log("llego al delete")
             // Construimos la URL para el DELETE con el correo como parámetro
             const url = `${DELETE}`;
 
-            try {
-                // Realizamos la solicitud DELETE al backend
-                const response = await axios.delete(url);
+            const confirmacion = window.confirm('¿Estás seguro que quieres eliminar tu cuenta?');
+            if (confirmacion) {
+                try {
+                    // Realizamos la solicitud DELETE al backend
+                    const response = await axios.delete(url);
 
-                // Verificamos si la eliminación fue exitosa
-                if (response.status === 204) {
-                    alert("Comercio eliminado con éxito");
-                    window.location.href = "index.html"; // Redirigir a otra página si es necesario
-                } else {
-                    throw new Error("Error en la eliminación");
+                    // Verificamos si la eliminación fue exitosa
+                    if (response.status === 204) {
+                        alert("Comercio eliminado con éxito");
+                        window.location.href = "../index.html"; // Redirigir a otra página si es necesario
+                    } else {
+                        throw new Error("Error en la eliminación");
+                    }
+
+                } catch (error) {
+                    console.error("Error al eliminar comercio:", error);
+                    alert("Hubo un problema con la eliminación del comercio.");
+                    window.location.href = "registroError.html"; // Redirigir en caso de error
                 }
-            } catch (error) {
-                console.error("Error al eliminar comercio:", error);
-                alert("Hubo un problema con la eliminación del comercio.");
-                window.location.href = "registroError.html"; // Redirigir en caso de error
-            }
-        }
+            }}
         ,
         async logoutComercio() {
             try {
-                // Realizamos la solicitud POST al backend para cerrar sesión
-                const response = await axios.post(LOGOUT);
+                console.log("logout 1");
 
-                // Verificamos si la respuesta fue exitosa
-                if (response.status === 200) {
-                    alert("Sesión cerrada con éxito");
-                    window.location.href = "index.html"; // Redirigir al login o a la página deseada
-                } else {
-                    throw new Error("Error al cerrar sesión");
+                const confirmacion = window.confirm('¿Estás seguro que quieres cerrar sesión?');
+                if (confirmacion) {
+                    // Realizamos la solicitud POST al backend para cerrar sesión
+                    const response = await axios.post(LOGOUT);
+
+                    // Verificamos si la respuesta fue exitosa
+                    if (response.status === 200) {
+                        alert("Sesión cerrada con éxito");
+                        window.location.href = "../index.html"; // Redirigir al login o a la página deseada
+                    } else {
+                        throw new Error("Error al cerrar sesión");
+                    }
                 }
             } catch (error) {
                 console.error("Error al cerrar sesión:", error);
                 alert("Hubo un problema al cerrar sesión.");
                 window.location.href = "registroError.html"; // Redirigir en caso de error
             }
+
+
         },
         async getActiveComercio() {
             await axios.get(RETRIEVE_ACTIVE)
